+++ tests/unit/plugins/module_utils/mock_info_api.py
@@ -46,93 +46,85 @@ class MockInfoApi:
 
     @staticmethod
     def get_security_config_response(response_type):
-        if response_type == 'api':
-            return [
-                {
-                    "id": "1"
-                }
-            ]
+        return [
+            {
+                "id": "1"
+            }
+        ]
 
     @staticmethod
     def get_certificate_response(response_type):
-        if response_type == 'api':
-            return [
-                {
-                    "id": "e940144f-393f-4e9c-8f54-9a4d57b38c48"
-                },
-                {
-                    "id": "4b6a46bc-b22c-47d8-b094-9a4dc1c6877d"
-                },
-                {
-                    "id": "37b76535-612b-456a-a694-1389f17632c7"
-                },
-                {
-                    "id": "e00b93aa-3a11-4242-9576-e18c0052c069"
-                },
-                {
-                    "id": "373dbe54-150f-47b9-8ce2-261307d05ebc"
-                },
-                {
-                    "id": "cedeb3ed-0203-405b-b6c0-b3117a331704"
-                }
-            ]
+        return [
+            {
+                "id": "e940144f-393f-4e9c-8f54-9a4d57b38c48"
+            },
+            {
+                "id": "4b6a46bc-b22c-47d8-b094-9a4dc1c6877d"
+            },
+            {
+                "id": "37b76535-612b-456a-a694-1389f17632c7"
+            },
+            {
+                "id": "e00b93aa-3a11-4242-9576-e18c0052c069"
+            },
+            {
+                "id": "373dbe54-150f-47b9-8ce2-261307d05ebc"
+            },
+            {
+                "id": "cedeb3ed-0203-405b-b6c0-b3117a331704"
+            }
+        ]
 
     @staticmethod
     def get_dns_response(response_type):
-        if response_type == 'api':
-            return [
-                {
-                    "id": "DNS1"
-                }
-            ]
+        return [
+            {
+                "id": "DNS1"
+            }
+        ]
 
     @staticmethod
     def get_ntp_response(response_type):
-        if response_type == 'api':
-            return [
-                {
-                    "id": "NTP1"
-                }
-            ]
+        return [
+            {
+                "id": "NTP1"
+            }
+        ]
 
     @staticmethod
     def get_one_id_response(response_type):
-        if response_type == 'api':
-            return [
-                {
-                    "id": "0"
-                }
-            ]
+        return [
+            {
+                "id": "0"
+            }
+        ]
 
     @staticmethod
     def get_email_destination_response(response_type):
-        if response_type == 'api':
-            return [
-                {
-                    "id": "9c3e5cba-17d5-4d64-b97c-350f91e2b714"
-                }
-            ]
+        return [
+            {
+                "id": "9c3e5cba-17d5-4d64-b97c-350f91e2b714"
+            }
+        ]
 
     @staticmethod
     def get_remote_support_contact_response(response_type):
-        if response_type == 'api':
-            return [
-                {
-                    "id": "0"
-                },
-                {
-                    "id": "1"
-                }
-            ]
+        return [
+            {
+                "id": "0"
+            },
+            {
+                "id": "1"
+            }
+        ]
 
     @staticmethod
     def get_remote_support_response(response_type):
-        if response_type == 'api':
-            return [
-                {
-                    "id": "0"
-                }
-            ]
+        return [
+            {
+                "id": "10"
+            }
+        ]
 
     @staticmethod
     def get_volumes_failed_msg():